api/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
python-dotenv==1.0.1
cohere==5.17.0
pydantic==2.10.4
python-multipart==0.0.20
requests==2.32.3
aiohttp==3.9.1
